+++ .flake8
@@ -1,0 +1,5 @@
+[flake8]
+
+ignore =
+    # Allow Black to set line length.
+    E501
