--- app/auth/templates/login.j2
@@ -1,14 +1,0 @@
-{% extends "base.j2" %}}
-
-{% import 'bootstrap/wtf.html' as wtf %}
-
-{% block app_content %}
-<div class="container">
-    <div class="row justify-content-center">
-        <div class="col-6">
-            <h1 class="text-muted">Please authenticate to proceed.</h1>
-            {{ wtf.quick_form(form, button_map={"submit": "success"}) }}
-        </div>
-    </div>
-</div>
-{% endblock %}
