+++ app/checks/forms.py
@@ -2,29 +2,28 @@ from flask_wtf import FlaskForm
 from wtforms import StringField, SubmitField
 from wtforms.validators import (
     DataRequired,
+    Length,
     ValidationError,
     URL,
 )
 
-from app.models import Check
+from app.checks.models import Check
 
 
-class CheckForm(FlaskForm):
-    """Base class for creating or editing a check."""
-
+class BaseCheckForm(FlaskForm):
     name = StringField("Name", validators=[DataRequired()])
-    url = StringField("URL", validators=[DataRequired(), URL()])
+    url = StringField("URL", validators=[DataRequired(), URL(), Length(max=2048)],)
 
 
-class CreateCheckForm(CheckForm):
+class CreateCheckForm(BaseCheckForm):
     submit = SubmitField("Create")
 
     def validate_name(self, name):
         if Check.query.filter_by(name=name.data).first() is not None:
-            raise ValidationError("Please use a unique check name.")
+            raise ValidationError("Please enter a unique name.")
 
 
-class EditCheckForm(CheckForm):
+class EditCheckForm(BaseCheckForm):
     submit = SubmitField("Save")
     cancel = SubmitField("Cancel")
