+++ app/checks/routes.py
@@ -1,15 +1,13 @@
 import arrow
 from flask import abort, current_app, flash, redirect, render_template, url_for
-from flask_login import login_required
 
 from app import db
-from app.models import Check, Event, Status
 from app.checks import bp
 from app.checks.forms import CreateCheckForm, DeleteCheckForm, EditCheckForm
+from app.checks.models import Check, Event, Status
 
 
 @bp.route("/checks", methods=["GET", "POST"])
-@login_required
 def checks():
     """Manage and create checks."""
     form = CreateCheckForm()
@@ -47,7 +45,6 @@ def view(id):
 
 
 @bp.route("/checks/edit/<id>", methods=["GET", "POST"])
-@login_required
 def edit(id):
     """Edit a check by ID."""
     check = Check.query.filter_by(id=id).first()
@@ -68,7 +65,6 @@ def edit(id):
 
 
 @bp.route("/checks/delete/<id>", methods=["GET", "POST"])
-@login_required
 def delete(id):
     """Delete a check by ID."""
     check = Check.query.filter_by(id=id).first()
