+++ app/dashboard/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+
+bp = Blueprint("dashboard", __name__, template_folder="templates")
+
+
+from app.dashboard import routes  # noqa
