+++ app/tasks/__init__.py
@@ -1,0 +1,2 @@
+from app.tasks.janitor import Janitor  # noqa
+from app.tasks.inquisitor import Inquisitor  # noqa
