+++ app/templates/base.j2
@@ -21,7 +21,7 @@
     </head>
     <body>
         <div class="container">
-            <a href="{{ url_for("main.dashboard") }}">Dashboard</a>
+            <a href="{{ url_for("dashboard.dashboard") }}">Dashboard</a>
             <a href="{{ url_for("checks.checks") }}">Checks</a>
 
             {{ render_messages() }}
