+++ pyproject.toml
@@ -11,7 +11,6 @@ flask = "^1.1.1"
 flask-wtf = "^0.14.2"
 flask-sqlalchemy = "^2.4.1"
 requests = "^2.22.0"
-flask-login = "^0.4.1"
 flask-migrate = "^2.5.2"
 flask-moment = "^0.9.0"
 psycopg2 = "^2.8.4"
